requirements.txt
cairocffi==1.7.1
CairoSVG==2.7.1
cffi==1.17.0
cssselect2==0.7.0
defusedxml==0.7.1
drawsvg==2.4.0
imageio==2.35.0
imageio-ffmpeg==0.5.1
numpy==1.26.4
pillow==10.4.0
pwkit==1.2.0
pycparser==2.22
PyYAML==6.0.2
setuptools==72.2.0
tinycss2==1.3.0
webencodings==0.5.1
